+++ layers/theme/app/components/global/Paragraph/Media.vue
@@ -6,7 +6,6 @@ import { normalizeDrupalMediaType } from '../../../utils/drupalMediaTypes'
 import { resolveResponsiveGridValue } from '../../../utils/responsiveGrid'
 import { useWindowSize } from '@vueuse/core'
 import { resolveStableMediaDeliveryProfile } from '../../../utils/imageDelivery'
-import { resolveHeadingTag } from '../../../utils/headingTag'
 import { resolveAlignClasses, type AlignConfig, type GridConfig } from '../../../utils/gridClasses'
 
 const props = defineProps<{
@@ -32,15 +31,12 @@ const props = defineProps<{
   }
 
   label?: string
-  header?: string
-  headerTag?: string
 
   editLink?: string
 }>()
 
 const resolvedWidth = computed(() => props.widthClass || props.width || '')
 const alignClasses = computed(() => resolveAlignClasses(props.align))
-const headingTag = computed(() => resolveHeadingTag(props.headerTag))
 const theme = useAppConfig().stirTheme
 const mediaHeightClass = computed(() =>
   theme.media.heights?.[props.mediaHeight || 'natural']
@@ -131,10 +127,6 @@ onMounted(() => {
     :parent-uuid="parentUuid"
   >
     <WrapDiv :align="alignClasses">
-      <component :is="headingTag" v-if="header">
-        {{ header }}
-      </component>
-
       <LazyUScrollArea
         v-if="props.masonry && hydrated"
         v-slot="{ item: node, index: i }"